Changed plugin namespace to comply with Poggit rules
